requirements.txt
fastapi 0.133.0
uvicorn 0.41.0
SQLAlchemy 2.0.46
psycopg2 2.9.11
python-dotenv 1.2.1




